Variables!
New error messages:
ERROR_UNKNOWN_VARIABLE, ERROR_BAD_LET_USAGE,
ERROR_BAD_ASSIGNMENT, ERROR_EQUALS_WITHOUT_LET

The system now uses maps to find whether a string is a token that already exists,
or if it is a new word that can become a variable. Eg, all single character operators
are in their own map and anytime one is come across, the current string before it
is stopped and evaluated as its own token.

Variables are now settable either as constants before hand, or as variables at
runtime. Using the "let" keyword allows variable assignment. As a consequence of this,
the ERROR_ILLEGAL_STR error is now redundant.

All spaces are replaced with '`' characters just for simplicity's sake.